Fixed an error in indexing addition to discretized observers, tested new index method, and tested/fixed weight windows and the weight window mesh class. Should be finished.

+++ packages/monte_carlo/event/population_control/src/MonteCarlo_PopulationControl.cpp
@@ -54,7 +54,7 @@ void PopulationControl::splitParticle( ParticleState& particle,
                                        ParticleBank& bank,
                                        unsigned number_of_particles ) const
 {
-  particle.multiplyWeight(1/number_of_particles);
+  particle.multiplyWeight(1/static_cast<double>(number_of_particles));
 
   this->pushSplitParticlesToBank(particle,
                                  bank,
